+++ examples/axis_inputs.rs
@@ -25,24 +25,21 @@ enum Action {
 struct Player;
 
 fn spawn_player(mut commands: Commands) {
+    // Describes how to convert from player inputs into those actions
+    let input_map = InputMap::default()
+        // Configure the left stick as a dual-axis
+        .insert(Action::Move, DualAxis::left_stick())
+        // Let's bind the right gamepad trigger to the throttle action
+        .insert(Action::Throttle, GamepadButtonType::RightTrigger2)
+        // And we'll use the right stick's x-axis as a rudder control
+        .insert(
+            // This will trigger if the axis is moved 10% or more in either direction.
+            Action::Rudder,
+            SingleAxis::symmetric(GamepadAxisType::RightStickX, 0.1),
+        )
+        .build();
     commands
-        .spawn(InputManagerBundle::<Action> {
-            // Stores "which actions are currently activated"
-            action_state: ActionState::default(),
-            // Describes how to convert from player inputs into those actions
-            input_map: InputMap::default()
-                // Configure the left stick as a dual-axis
-                .insert(Action::Move, DualAxis::left_stick())
-                // Let's bind the right gamepad trigger to the throttle action
-                .insert(Action::Throttle, GamepadButtonType::RightTrigger2)
-                // And we'll use the right stick's x-axis as a rudder control
-                .insert(
-                    // This will trigger if the axis is moved 10% or more in either direction.
-                    Action::Rudder,
-                    SingleAxis::symmetric(GamepadAxisType::RightStickX, 0.1),
-                )
-                .build(),
-        })
+        .spawn(InputManagerBundle::with_map(input_map))
         .insert(Player);
 }
 
@@ -62,8 +59,8 @@ fn move_player(query: Query<&ActionState<Action>, With<Player>>) {
 
     if action_state.pressed(&Action::Throttle) {
         // Note that some gamepad buttons are also tied to axes, so even though we used a
-        // GamepadbuttonType::RightTrigger2 binding to trigger the throttle action, we can get a
-        // variable value here if you have a variable right trigger on your gamepad.
+        // GamepadButtonType::RightTrigger2 binding to trigger the throttle action,
+        // we can get a variable value here if you have a variable right trigger on your gamepad.
         //
         // If you don't have a variable trigger, this will just return 0.0 when not pressed and 1.0
         // when pressed.